requirements.txt
#pip install -r requirements.txt
colorama
python-dotenv
pydantic-ai
streamlit
altair